refactor(validation): rename validation module to evaluation
