+++ docs/evaluation/calculate_base_metrics.md
@@ -1,0 +1,3 @@
+# Calculate base metrics
+
+::: eis_toolkit.evaluation.calculate_base_metrics
