+++ docs/evaluation/classification_label_evaluation.md
@@ -1,0 +1,3 @@
+# Classification label evaluation
+
+::: eis_toolkit.evaluation.classification_label_evaluation
