+++ docs/evaluation/classification_probability_evaluation.md
@@ -1,0 +1,3 @@
+# Classification probability evaluation
+
+::: eis_toolkit.evaluation.classification_probability_evaluation
