+++ docs/evaluation/plot_confusion_matrix.md
@@ -1,0 +1,3 @@
+# Plot confusion matrix
+
+::: eis_toolkit.evaluation.plot_confusion_matrix
