+++ docs/evaluation/plot_nn_model_performance.md
@@ -1,3 +1,3 @@
 # Plot neural network training performance (accuracy and loss)
 
-::: eis_toolkit.validation.plot_nn_model_performance
+::: eis_toolkit.evaluation.plot_nn_model_performance
