+++ docs/evaluation/plot_prediction_area_curves.md
@@ -1,0 +1,3 @@
+# Plot prediction-area (P-A) curves
+
+::: eis_toolkit.evaluation.plot_prediction_area_curves
