+++ docs/evaluation/plot_rate_curve.md
@@ -1,0 +1,3 @@
+# Plot rate curve
+
+::: eis_toolkit.evaluation.plot_rate_curve
