--- docs/validation/calculate_base_metrics.md
@@ -1,3 +1,0 @@
-# Calculate base metrics
-
-::: eis_toolkit.validation.calculate_base_metrics
