--- docs/validation/classification_label_evaluation.md
@@ -1,3 +1,0 @@
-# Classification label evaluation
-
-::: eis_toolkit.validation.classification_label_evaluation
