--- docs/validation/classification_probability_evaluation.md
@@ -1,3 +1,0 @@
-# Classification probability evaluation
-
-::: eis_toolkit.validation.classification_probability_evaluation
