--- docs/validation/plot_confusion_matrix.md
@@ -1,3 +1,0 @@
-# Plot confusion matrix
-
-::: eis_toolkit.validation.plot_confusion_matrix
