--- docs/validation/plot_prediction_area_curves.md
@@ -1,3 +1,0 @@
-# Plot prediction-area (P-A) curves
-
-::: eis_toolkit.validation.plot_prediction_area_curves
