--- docs/validation/plot_rate_curve.md
@@ -1,3 +1,0 @@
-# Plot rate curve
-
-::: eis_toolkit.validation.plot_rate_curve
