+++ eis_toolkit/evaluation/plot_rate_curve.py
